+++ .changelog/2836.txt
@@ -1,3 +1,3 @@
 ```release-note:enhancement
-resource/mongodbatlas_advanced_cluster: Ensures asymmetric auto-scaling is not defined in the cluster when using the old sharding configuration
+resource/mongodbatlas_advanced_cluster: Adjusts create operation to support cluster tier auto scaling per shard.
 ```
